Fixed a bug where after headset was plugged in once it would auto start again
If the switch was thrown and no headset was present the alert would
show. If a headset was then inserted the alert would disappear as it
should. If the headset was removed and then reinserted the switch would
be thrown without user interaction. This was fixed by stopping the
timer for headset checking in the alert button callback.